Tooltip top position is now better calculated
It wasn't down far enough when logged in on links that had a line break. It was also too far down when logged out due to the absence of the admin bar. Now the height of the link is added to its top offset to get the starting point -- as it should've been all along.