Excerpts are now the first paragraph regardless of its length
This is overly simplistic for the time being and requires that you don't have non-paragraph elements before the first paragraph. If that first paragraph is stupidly short, so too will be the post's excerpt.

The correct play is probably to use WP's new(ish) HTML processor to fish out that first paragraph regardless of what comes before. I forgot it existed until writing this commit.