Tooltip positioning delayed until first hover
This should (finally) clear up the issue where slow loading of images or embeds would set these "coordinates" too soon.

See, what had happened was, is once the other content loaded and moved the link downward, the tooltip's position would be wrong. It could show way above the link or, as it did in most cases, appear directly under the cursor. Well, if you're no longer hovering the link the tooltip goes away, so the browser would go into a loop trying to show and then hide the tooltip repeatedly. The user would see nothing by their cursor not turning into a pointer over a link like it should.

I've known this was an issue for awhile but kept procrastinating on a fix. Still need to clean the code up a little, but this is finally fixed. Hopefully.